add create_python_mapping
Summary: This diff adds argparse to the TritonParse project. The code changes include adding the argparse library to the project and modifying the existing code to use argparse for command line argument parsing. The get_file_extension function is also added to extract the file extension from a given filename. The create_python_mapping function is modified to accept a list of tuples containing the IR maps and returns a dictionary of the mapping.

Reviewed By: davidberard98

Differential Revision: D74284154

fbshipit-source-id: a7b5279d4f85387317cea2c138b5b3e9bb2a32ac